Change fonts
User can now choose what font they'd like to write their entry in. DearDiary updates font in real-time. Chosen font is saved and stored in the file that's created for the entry that is later used by show_selected_entry to display entry in whatever font it was written in. Font is maintained when editing an entry as well.