Fresh start and new look
Learned more about tkinter, ttk, and ttkbootstrap. Currently have a functional diary app that displays the date, has an entry for the user to enter a title, and a scrolled text box for writing entries. Upon save, entry is then saved to a file within the created (upon initialization) directory 'Entries'. As of this commit, user can only write and save a diary entry. More features to be added.