View and Edit Entry
User can now view existing entries and edit them if they wish. Upon editing an entry the original title cannot be altered to avoid errors in overwriting the existing file. When changes are made and the edit entry button is clicked, the file is overwritten with an additional line of text declaring that the file was edited and the date and time it was last edited. If user only want to view entry and return to creating a new entry they can clear the text box and title by clicking Create New Entry button.